chaostooling-extension-db/pyproject.toml
[project]
name = "chaostoolkit-extension-db"
dynamic = ["version"]
description = "Chaos Toolkit extension for connectivity probes/actions across PostgreSQL, MySQL, MSSQL, MongoDB, Redis, Cassandra, Kafka, RabbitMQ, and ActiveMQ"
authors = [{name = "Morgan Wigge", email = "morgan@wigge.nu"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "chaostoolkit-lib>=1.42.1",

    # SQL/NoSQL databases
    "psycopg2-binary>=2.9.0",           # PostgreSQL
    "mysql-connector-python>=8.0.0",    # MySQL
    "pyodbc>=4.0.39",                   # MSSQL (requires system unixodbc-dev & MS ODBC driver 18)
    "pymongo>=4.0.0",                   # MongoDB
    "redis>=4.5.0",                     # Redis
    "cassandra-driver>=3.28.0",         # Cassandra

    # Messaging systems
    "kafka-python>=2.0.2",              # Kafka
    "pika>=1.3.0",                      # RabbitMQ
    "stomp.py>=8.1.0",                  # ActiveMQ (STOMP support enabled on broker)

    "requests>=2.31.0",
    "psutil>=5.9.0",                    # Process management for disaster recovery tests
    # Observability via chaosotel (chaostooling-otel)
    "chaostooling-otel>=1.0.0",
    # "chaosreporting>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "mypy>=1.19.0",
    "black>=25.11.0",
    "ruff>=0.14.7",
    "isort",
    "bandit>=1.7.0",
    "codespell>=2.3.0",
]

[build-system]
requires = ["setuptools", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "chaosdb/_version.py"
fallback_version = "0.1.0"

[tool.setuptools.packages.find]
include = ["chaosdb*"]

[tool.pdm.scripts]
lint = {composite = ["ruff check chaosdb/", "isort --check-only --profile black chaosdb/", "black --check --diff chaosdb/", "mypy chaosdb/", "bandit -r chaosdb", "codespell chaosdb/"]}
format = {composite = ["isort --profile black chaosdb/", "black chaosdb/", "ruff format chaosdb/ --fix"]}
test = {cmd = "pytest"}

[tool.black]
color = true
line-length = 80
target-version = ['py310']

[tool.mypy]
namespace_packages = true
python_version = "3.10"

[tool.ruff]
line-length = 80
target-version = "py310"
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "build",
    "dist",
    "chaostoolkit_extension_db.egg-info",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[tool.isort]
line_length = 80
known_first_party = 'chaosdb'
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.codespell]
ignore_words_list = [
    "chaosdb",
    "chaosotel",
    "chaostoolkit",
    "postgres",
    "postgresql",
    "mongodb",
    "activemq",
    "rabbitmq",
    "mssql",
]
skip = "*.pyc,*.pyo,*.egg-info,.git,.venv,__pycache__"
